Moved all docs, examples to libdogleg-doc
This was done because libdogleg-dev (package that contained these previously) is
now multi-arch:same. Since some of the documentation is generated by other
tools, there were minor differences due mostly to different versions of those
tools. Since those files were no longer identical, the package couldn't be
multi-arch:same. The new package (libdogleg-doc) is arch:all, so this ceases to
be a problem.