reworked outlier business to work on measurement GROUPS
Often I have clusters of measurements that should be outlier-rejected as a
group. Example: I have pixel features that produce an x error and a y error. I
want to reject FEATURES, so I want to operate on pairs of measurements as a
block.

This patch is somewhat experimental still. If nothing else, I need many more
comments explaining what the measurements and groups are